requirements.txt
google-adk
requests
beautifulsoup4
pydantic
google-genai
sqlalchemy>=2.0.0
alembic>=1.12.0
aiohttp
asyncio
